trace: Implement trace library abstraction
 - implement module to allow platform specific trace library
 - move current trace implementation into the module
