+++ app/iscsi_top/iscsi_top.cpp
@@ -50,6 +50,7 @@
 
 extern "C" {
 #include "spdk/trace.h"
+#include "spdk_internal/trace_shm.h"
 #include "iscsi/conn.h"
 }
